feat: Add asset images, set up file structure, test API fetch, basic structural styling for App component
